Move `CairoPluginProjectBuilder` to `utils/scarb-test-support/`
